pyproject.toml
[tool.poetry]
name = "faust-demo"
version = "0.1.0"
description = ""
authors = ["Philip Adzanoukpe <epigos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
faust = {extras = ["fast", "debug"], version = "^1.10.4"}
boto3 = "^1.17.28"
python-schema-registry-client = "^1.8.1"
dataclasses-avroschema = "^0.20.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
